+++ tests/test_order.py
@@ -1,14 +1,15 @@
-import copy
-from typing import Any, Dict
 from unittest import mock
 
 import pytest
 import requests
+import requests_mock as req_mock
 
 from up42 import asset, order
 
 from .fixtures import fixtures_globals as constants
-from .fixtures import fixtures_order
+
+ORDER_URL = f"{constants.API_HOST}/v2/orders/{constants.ORDER_ID}"
+ORDER_PLACEMENT_URL = f"{constants.API_HOST}/v2/orders?workspaceId={constants.WORKSPACE_ID}"
 
 
 @pytest.fixture(autouse=True)
@@ -19,201 +20,173 @@ def workspace():
         yield
 
 
-def test_init(order_mock):
-    assert isinstance(order_mock, order.Order)
-    assert order_mock.order_id == constants.ORDER_ID
+class TestOrder:
+    @pytest.fixture(scope="class", params=["catalog", "tasking"])
+    def order_parameters(self, request) -> order.OrderParams:
+        geometry_key = "aoi" if request.param == "catalog" else "geometry"
+        return {
+            "dataProduct": "some-data-product",
+            "params": {geometry_key: {"some": "shape"}},
+        }
 
+    def test_should_initialize_with_info_provided(self):
+        info = {"some": "data"}
+        order_obj = order.Order(constants.ORDER_ID, info)
+        assert order_obj.order_id == constants.ORDER_ID
+        assert order_obj._info == info  # pylint: disable=protected-access
 
-def test_order_info(order_mock):
-    assert order_mock.info
-    assert order_mock.info["id"] == constants.ORDER_ID
+    def test_should_initialize(self, requests_mock: req_mock.Mocker):
+        info = {"some": "data"}
+        requests_mock.get(url=ORDER_URL, json=info)
+        order_obj = order.Order(constants.ORDER_ID)
+        assert order_obj.order_id == constants.ORDER_ID
+        assert order_obj.info == info
 
+    def test_should_provide_representation(self):
+        info = {
+            "id": constants.ORDER_ID,
+            "status": "PLACED",
+            "createdAt": "2023-01-01T12:00:00Z",
+            "updatedAt": "2023-01-01T12:30:00Z",
+        }
+        order_obj = order.Order(order_id=constants.ORDER_ID, order_info=info)
+        expected_repr = (
+            f"Order(order_id: {constants.ORDER_ID}, status: PLACED,"
+            "createdAt: 2023-01-01T12:00:00Z, updatedAt: 2023-01-01T12:30:00Z)"
+        )
+        assert repr(order_obj) == expected_repr
 
-def test_repr():
-    order_info = {
-        "id": "your_order_id",
-        "status": "PLACED",
-        "createdAt": "2023-01-01T12:00:00Z",
-        "updatedAt": "2023-01-01T12:30:00Z",
-    }
-    order_placed = order.Order(order_id="your_order_id", order_info=order_info)
+    def test_should_provide_status(self, requests_mock: req_mock.Mocker):
+        info = {"status": "random"}
+        requests_mock.get(url=ORDER_URL, json=info)
+        order_obj = order.Order(order_id=constants.ORDER_ID, order_info=info)
+        assert order_obj.status == info["status"]
 
-    expected_repr = (
-        "Order(order_id: your_order_id, status: PLACED,"
-        "createdAt: 2023-01-01T12:00:00Z, updatedAt: 2023-01-01T12:30:00Z)"
+    @pytest.mark.parametrize(
+        "info, expected",
+        [
+            ({"type": "ARCHIVE"}, {}),
+            (
+                {"type": "TASKING", "orderDetails": {"order": "details"}},
+                {"order": "details"},
+            ),
+        ],
+        ids=["ARCHIVE", "TASKING"],
     )
-    assert repr(order_placed) == expected_repr
+    def test_should_provide_order_details(self, requests_mock: req_mock.Mocker, info: dict, expected: dict):
+        requests_mock.get(url=ORDER_URL, json=info)
+        order_obj = order.Order(order_id=constants.ORDER_ID, order_info=info)
+        assert order_obj.order_details == expected
 
+    @pytest.mark.parametrize(
+        "info, expected",
+        [
+            ({"status": "FULFILLED"}, True),
+            ({"status": "OTHER STATUS"}, False),
+        ],
+        ids=["FULFILLED", "OTHER STATUS"],
+    )
+    def test_should_compute_is_fulfilled(self, requests_mock: req_mock.Mocker, info: dict, expected: bool):
+        requests_mock.get(url=ORDER_URL, json=info)
+        order_obj = order.Order(order_id=constants.ORDER_ID, order_info=info)
+        assert order_obj.is_fulfilled == expected
 
-@pytest.mark.parametrize("status", ["PLACED", "FULFILLED"])
-def test_order_status(order_mock, status, monkeypatch):
-    monkeypatch.setattr(order.Order, "info", {"status": status})
-    assert order_mock.status == status
+    def test_should_get_assets_if_fulfilled(self, requests_mock: req_mock.Mocker):
+        info = {"id": constants.ORDER_ID, "status": "FULFILLED"}
+        requests_mock.get(url=ORDER_URL, json=info)
+        url_asset_info = f"{constants.API_HOST}/v2/assets?search={constants.ORDER_ID}"
+        asset_info = {
+            "content": [
+                {
+                    "id": constants.ASSET_ORDER_ID,
+                }
+            ],
+            "totalPages": 1,
+            "totalElements": 1,
+        }
+        requests_mock.get(url=url_asset_info, json=asset_info)
+        order_obj = order.Order(order_id=constants.ORDER_ID)
+        (asset_obj,) = order_obj.get_assets()
+        assert isinstance(asset_obj, asset.Asset)
+        assert asset_obj.asset_id == constants.ASSET_ORDER_ID
 
+    @pytest.mark.parametrize(
+        "status",
+        [
+            "CREATED",
+            "BEING_PLACED",
+            "PLACED",
+            "PLACEMENT_FAILED",
+            "DELIVERY_INITIALIZATION_FAILED",
+            "DOWNLOAD_FAILED",
+            "DOWNLOADED",
+            "FAILED_PERMANENTLY",
+        ],
+    )
+    def test_fails_to_get_assets_if_not_fulfilled(self, requests_mock: req_mock.Mocker, status: str):
+        info = {"status": status}
+        requests_mock.get(url=ORDER_URL, json=info)
+        order_obj = order.Order(order_id=constants.ORDER_ID, order_info=info)
+        with pytest.raises(order.UnfulfilledOrder, match=f".*{constants.ORDER_ID}.*{status}"):
+            order_obj.get_assets()
 
-@pytest.mark.parametrize(
-    "status, order_type, order_details",
-    [
-        ("PLACED", "TASKING", {"subStatus": "FEASIBILITY_WAITING_UPLOAD"}),
-        ("FULFILLED", "ARCHIVE", {}),
-    ],
-)
-def test_order_details(order_mock, status, order_type, order_details, monkeypatch):
-    monkeypatch.setattr(
-        order.Order,
+    @pytest.mark.parametrize(
         "info",
-        {"status": status, "type": order_type, "orderDetails": order_details},
+        [{"type": "ARCHIVE"}, {"type": "TASKING", "orderDetails": {"subStatus": "substatus"}}],
+        ids=["ARCHIVE", "TASKING"],
     )
-    assert order_mock.order_details == order_details
+    def test_should_track_order_status_until_fulfilled(self, requests_mock: req_mock.Mocker, info: dict):
+        responses = []
+        statuses = ["PLACED", "BEING_FULFILLED", "FULFILLED"]
+        responses = [{"json": {"status": status, **info}} for status in statuses]
+        requests_mock.get(ORDER_URL, responses)
+        order_obj = order.Order(order_id=constants.ORDER_ID)
+        assert order_obj.track_status(report_time=0.1) == "FULFILLED"
 
+    @pytest.mark.parametrize("status", ["FAILED", "FAILED_PERMANENTLY"])
+    def test_fails_to_track_order_if_status_not_valid(self, requests_mock: req_mock.Mocker, status: str):
+        info = {"status": status, "type": "ANY"}
+        requests_mock.get(
+            url=ORDER_URL,
+            json=info,
+        )
+        order_obj = order.Order(order_id=constants.ORDER_ID)
+        with pytest.raises(order.FailedOrder):
+            order_obj.track_status()
 
-@pytest.mark.parametrize(
-    "status,expected",
-    [("NOT STARTED", False), ("PLACED", False), ("FULFILLED", True)],
-)
-def test_is_fulfilled(order_mock, status, expected, monkeypatch):
-    monkeypatch.setattr(order.Order, "info", {"status": status})
-    assert order_mock.is_fulfilled == expected
+    def test_should_estimate(self, auth_mock, requests_mock: req_mock.Mocker, order_parameters: order.OrderParams):
+        order_estimate_url = f"{constants.API_HOST}/v2/orders/estimate"
+        expected_credits = 100
+        requests_mock.post(
+            url=order_estimate_url,
+            json={
+                "summary": {"totalCredits": expected_credits},
+                "errors": [],
+            },
+        )
+        assert order.Order.estimate(auth_mock, order_parameters) == expected_credits
 
+    def test_should_place_order(self, requests_mock: req_mock.Mocker, order_parameters: order.OrderParams):
+        info = {"status": "SOME STATUS"}
+        requests_mock.get(
+            url=ORDER_URL,
+            json=info,
+        )
+        requests_mock.post(
+            url=ORDER_PLACEMENT_URL,
+            json={"results": [{"id": constants.ORDER_ID}], "errors": []},
+        )
+        order_obj = order.Order.place(order_parameters, constants.WORKSPACE_ID)
+        assert order_obj == order.Order(order_id=constants.ORDER_ID, order_info=info)
 
-def test_order_parameters(order_mock):
-    assert not order_mock.order_parameters
-
-
-@pytest.mark.parametrize(
-    "status",
-    [
-        "FULFILLED",
-        "BEING_FULFILLED",
-    ],
-)
-def test_get_assets_should_search_assets_by_order_id(requests_mock, status):
-    order_response = {"id": constants.ORDER_ID, "status": status}
-    url_order_info = f"{constants.API_HOST}/v2/orders/{constants.ORDER_ID}"
-    requests_mock.get(url=url_order_info, json=order_response)
-    response_assets_json: Dict[str, Any] = copy.deepcopy(fixtures_order.JSON_GET_ASSETS_RESPONSE)
-    response_assets_json["totalPages"] = 2
-    url_asset_info = f"{constants.API_HOST}/v2/assets?search={constants.ORDER_ID}&page=0"
-    requests_mock.get(url=url_asset_info, json=response_assets_json)
-    response_assets_json["pageable"]["pageNumber"] = 2
-    url_asset_info = f"{constants.API_HOST}/v2/assets?search={constants.ORDER_ID}&page=1"
-    requests_mock.get(url=url_asset_info, json=response_assets_json)
-    order_placed = order.Order(order_id=constants.ORDER_ID)
-    asset_returned = list(order_placed.get_assets())
-    assert all(isinstance(asset_element, asset.Asset) for asset_element in asset_returned)
-    assert all(asset_element.asset_id == constants.ASSET_ORDER_ID for asset_element in asset_returned)
-
-
-@pytest.mark.parametrize(
-    "status",
-    [
-        "CREATED",
-        "BEING_PLACED",
-        "PLACED",
-        "PLACEMENT_FAILED",
-        "DELIVERY_INITIALIZATION_FAILED",
-        "DOWNLOAD_FAILED",
-        "DOWNLOADED",
-        "FAILED_PERMANENTLY",
-    ],
-)
-def test_should_fail_to_get_assets_for_unfulfilled_order(requests_mock, status):
-    order_response = {"id": constants.ORDER_ID, "status": status}
-    url_order_info = f"{constants.API_HOST}/v2/orders/{constants.ORDER_ID}"
-    requests_mock.get(url=url_order_info, json=order_response)
-    order_placed = order.Order(order_id=constants.ORDER_ID)
-    with pytest.raises(ValueError):
-        list(order_placed.get_assets())
-
-
-def test_place_order(catalog_order_parameters, order_mock, requests_mock):
-    requests_mock.post(
-        url=f"{constants.API_HOST}/v2/orders?workspaceId={constants.WORKSPACE_ID}",
-        json={
-            "results": [{"index": 0, "id": constants.ORDER_ID}],
-            "errors": [],
-        },
-    )
-    order_placed = order.Order.place(catalog_order_parameters, constants.WORKSPACE_ID)
-    assert order_placed == order_mock
-    assert order_placed.order_id == constants.ORDER_ID
-    assert order_placed.order_parameters == catalog_order_parameters
-
-
-def test_place_order_fails_if_response_contains_error(catalog_order_parameters, requests_mock):
-    error_content = "test error"
-    requests_mock.post(
-        url=f"{constants.API_HOST}/v2/orders?workspaceId={constants.WORKSPACE_ID}",
-        json={
-            "results": [],
-            "errors": [{"message": error_content}],
-        },
-    )
-    with pytest.raises(ValueError) as err:
-        order.Order.place(catalog_order_parameters, constants.WORKSPACE_ID)
-    assert error_content in str(err.value)
-
-
-def test_track_status_running(requests_mock):
-    url_job_info = f"{constants.API_HOST}/v2/orders/{constants.ORDER_ID}"
-    status_responses = [
-        {
-            "json": {
-                "status": "PLACED",
-                "type": "TASKING",
-                "orderDetails": {"subStatus": "NON_STANDARD_STATUS"},
-            }
-        },
-        {
-            "json": {
-                "status": "BEING_FULFILLED",
-                "type": "TASKING",
-                "orderDetails": {"subStatus": "FEASIBILITY_WAITING_UPLOAD"},
-            }
-        },
-        {
-            "json": {
-                "status": "FULFILLED",
-                "type": "TASKING",
-                "orderDetails": {"subStatus": "FEASIBILITY_WAITING_UPLOAD"},
-            }
-        },
-    ]
-    requests_mock.get(url_job_info, status_responses)
-    order_test = order.Order(constants.ORDER_ID)
-    order_status = order_test.track_status(report_time=1)
-    assert order_status == "FULFILLED"
-
-
-@pytest.mark.parametrize("status", ["FULFILLED"])
-def test_track_status_pass(order_mock, status, requests_mock):
-    url_job_info = f"{constants.API_HOST}/v2/orders/{order_mock.order_id}"
-    requests_mock.get(url=url_job_info, json={"status": status})
-
-    order_status = order_mock.track_status()
-    assert order_status == status
-
-
-@pytest.mark.parametrize("status", ["FAILED", "FAILED_PERMANENTLY"])
-def test_track_status_fail(order_mock, status, requests_mock):
-    url_job_info = f"{constants.API_HOST}/v2/orders/{order_mock.order_id}"
-    requests_mock.get(
-        url=url_job_info,
-        json={"status": status, "type": "ARCHIVE"},
-    )
-
-    with pytest.raises(ValueError):
-        order_mock.track_status()
-
-
-def test_estimate_order(catalog_order_parameters, auth_mock, requests_mock):
-    expected_payload = {
-        "summary": {"totalCredits": 100, "totalSize": 0.1, "unit": "SQ_KM"},
-        "results": [{"index": 0, "credits": 100, "unit": "SQ_KM", "size": 0.1}],
-        "errors": [],
-    }
-    url_order_estimation = f"{constants.API_HOST}/v2/orders/estimate"
-    requests_mock.post(url=url_order_estimation, json=expected_payload)
-    estimation = order.Order.estimate(auth_mock, catalog_order_parameters)
-    assert isinstance(estimation, int)
-    assert estimation == 100
+    def test_fails_to_place_order_if_response_contains_error(
+        self, requests_mock: req_mock.Mocker, order_parameters: order.OrderParams
+    ):
+        error_msg = "test error"
+        order_response_with_error = {"results": [], "errors": [{"message": error_msg}]}
+        requests_mock.post(
+            url=ORDER_PLACEMENT_URL,
+            json=order_response_with_error,
+        )
+        with pytest.raises(order.FailedOrderPlacement, match=error_msg):
+            order.Order.place(order_parameters, constants.WORKSPACE_ID)