+++ up42/order.py
@@ -54,7 +54,19 @@ OrderSubtatus = Literal[
 ]
 
 
-class OrderParams(TypedDict):
+class UnfulfilledOrder(ValueError):
+    pass
+
+
+class FailedOrder(ValueError):
+    pass
+
+
+class FailedOrderPlacement(ValueError):
+    pass
+
+
+class OrderParams(TypedDict, total=False):
     """
     Represents the stucture data format for the order parameters.
     dataProduct: The dataProduct id for the specific product configuration.
@@ -84,11 +96,9 @@ class Order:
     def __init__(
         self,
         order_id: str,
-        order_parameters: Optional[dict] = None,
         order_info: Optional[dict] = None,
     ):
         self.order_id = order_id
-        self.order_parameters = order_parameters
         if order_info is not None:
             self._info = order_info
         else:
@@ -127,8 +137,7 @@ class Order:
         Gets the Order Details. Only for tasking type orders, archive types return empty.
         """
         if self.info["type"] == "TASKING":
-            order_details = self.info["orderDetails"]
-            return order_details
+            return self.info["orderDetails"]
         logger.info("Order is not TASKING type. Order details are not provided.")
         return {}
 
@@ -144,6 +153,7 @@ class Order:
         """
         Gets the Order assets or results.
         """
+        current_info = copy.deepcopy(self._info)
 
         def get_pages(endpoint: str, params: dict[str, Any]):
             response = self.session.get(host.endpoint(endpoint), params=params).json()
@@ -155,15 +165,15 @@ class Order:
                     break
                 response = self.session.get(host.endpoint(endpoint), params=params).json()
 
-        if not (self.is_fulfilled) and self.status != "BEING_FULFILLED":
-            raise ValueError(f"Order {self.order_id} is not valid. Current status is {self.status}")
+        if current_info["status"] != "FULFILLED" and current_info["status"] != "BEING_FULFILLED":
+            raise UnfulfilledOrder(f"Order {self.order_id} is not valid. Current status is {current_info['status']}")
         params = {"search": self.order_id, "page": 0}
         for page in get_pages("/v2/assets", params):
             for asset_info in page:
                 yield asset.Asset(base.workspace.auth, asset_info=asset_info)
 
     @classmethod
-    def place(cls, order_parameters: dict, workspace_id: str) -> "Order":
+    def place(cls, order_parameters: OrderParams, workspace_id: str) -> "Order":
         """
         Places an order.
 
@@ -178,9 +188,9 @@ class Order:
         response_json = cls.session.post(url=url, json=_translate_construct_parameters(order_parameters)).json()
         if response_json["errors"]:
             message = response_json["errors"][0]["message"]
-            raise ValueError(f"Order was not placed: {message}")
+            raise FailedOrderPlacement(f"Order was not placed: {message}")
         order_id = response_json["results"][0]["id"]
-        order = cls(order_id=order_id, order_parameters=order_parameters)
+        order = cls(order_id=order_id)
         logger.info("Order %s is now %s.", order.order_id, order.status)
         return order
 
@@ -211,7 +221,7 @@ class Order:
         )
         return estimated_credits
 
-    def track_status(self, report_time: int = 120) -> str:
+    def track_status(self, report_time: float = 120) -> str:
         """
         Continuously gets the order status until order is fulfilled or failed.
 
@@ -229,42 +239,19 @@ class Order:
         Returns:
             str: The final order status.
         """
-
-        def substatus_messages(substatus: str) -> str:
-            substatus_user_messages = {
-                "FEASIBILITY_WAITING_UPLOAD": "Wait for feasibility.",
-                "FEASIBILITY_WAITING_RESPONSE": "Feasibility is ready.",
-                "QUOTATION_WAITING_UPLOAD": "Wait for quotation.",
-                "QUOTATION_WAITING_RESPONSE": "Quotation is ready",
-                "QUOTATION_ACCEPTED": "In progress.",
-            }
-
-            if substatus in substatus_user_messages:
-                message = substatus_user_messages[substatus]
-                return f"{substatus}, {message}"
-            return f"{substatus}"
-
         logger.info("Tracking order status, reporting every %s seconds...", report_time)
-        time_asleep = 0
-
-        # check order details and react for tasking orders.
-
-        while not self.is_fulfilled:
-            status = self.status
-            substatus_message = (
-                substatus_messages(self.order_details.get("subStatus", "")) if self.info["type"] == "TASKING" else ""
-            )
-            if status in ["PLACED", "BEING_FULFILLED"]:
-                if time_asleep != 0 and time_asleep % report_time == 0:
-                    logger.info("Order is %s! - %s", status, self.order_id)
-                    logger.info(substatus_message)
-
-            elif status in ["FAILED", "FAILED_PERMANENTLY"]:
+        time_asleep: float = 0
+        current_info = copy.deepcopy(self._info)
+        while (status := current_info["status"]) != "FULFILLED":
+            sub_status = current_info.get("orderDetails", {}).get("subStatus")
+            status += f": {sub_status}" if sub_status is not None else ""
+            if time_asleep != 0 and time_asleep % report_time == 0:
                 logger.info("Order is %s! - %s", status, self.order_id)
-                raise ValueError("Order has failed!")
-
+            if status in ["FAILED", "FAILED_PERMANENTLY"]:
+                raise FailedOrder("Order has failed!")
             time.sleep(report_time)
             time_asleep += report_time
+            current_info = copy.deepcopy(self.info)
 
         logger.info("Order is fulfilled successfully! - %s", self.order_id)
-        return self.status
+        return current_info["status"]