+++ CHANGELOG.md
@@ -30,11 +30,21 @@ You can check your current version with the following command:
 
 For more information, see [UP42 Python package description](https://pypi.org/project/up42-py/).
 
+## 2.1.0a3
+
+**Sep 3, 2024**
+- `Order.get_assets` now allows to get assets from orders in `BEING_FULFILLED` state.
+
+## 2.1.0a2
+
+**Sep 2, 2024**
+- Fix test coverage for `Order::estimate`, `Order::place`, and `Order::track_status` methods.
+- Change `Order::track_status` report_time input type to float.
+
 ## 2.1.0a1
 
-**Aug 26, 2024**
-- `Order.get_assets` now allows to get assets from orders in `BEING_FULFILLED` state.
-- `Order.get_assets` now returning an iterator instear of list.
+**Aug 30, 2024**
+- Improve test coverage `Order` class.
 
 ## 2.0.1